+++ src/controllers/Game.js
@@ -5,10 +5,13 @@ const Player = require("./Player");
 const GameHistory = require("./GameHistory");
 const { generateError, sendError } = require("../helpers/functions/Error");
 const { checkRequestData } = require("../helpers/functions/Express");
-const { isVillagerSideAlive, isWerewolfSideAlive, areAllPlayersDead, isGameDone, getPlayerWithAttribute } = require("../helpers/functions/Game");
-const { populate: fullGamePopulate, turnPreNightActionsOrder, turnNightActionsOrder } = require("../helpers/constants/Game");
-const { groupNames } = require("../helpers/constants/Role");
-const { getPlayerRoles } = require("../helpers/functions/Role");
+const {
+    isVillagerSideAlive, isWerewolfSideAlive, areAllPlayersDead, getPlayersWithAttribute, getPlayersWithRole, getGameTurNightActionsOrder,
+    areLoversTheOnlyAlive, isGameDone, getPlayerWithRole, getPlayersWithSide, areAllWerewolvesAlive, getAlivePlayers,
+} = require("../helpers/functions/Game");
+const { getPlayerAttribute } = require("../helpers/functions/Player");
+const { getRoles, getGroupNames } = require("../helpers/functions/Role");
+const { populate: fullGamePopulate } = require("../helpers/constants/Game");
 const { filterOutHTMLTags } = require("../helpers/functions/String");
 
 exports.getFindPopulate = projection => {
@@ -40,30 +43,38 @@ exports.findOne = async(search, projection, options = {}) => {
     return game;
 };
 
-exports.fillFirstWaiting = data => {
-    const firstWaiting = { for: "all", to: "elect-sheriff" };
-    data.waiting = [firstWaiting];
-};
-
 exports.checkUserCurrentGames = async userId => {
     if (await Game.countDocuments({ gameMaster: userId, status: "playing" })) {
         throw generateError("GAME_MASTER_HAS_ON_GOING_GAMES", "The game master has already game with status `playing`.");
     }
 };
 
+exports.fillTickData = game => {
+    game.tick = 1;
+    game.phase = "night";
+    game.turn = 1;
+};
+
 exports.checkRolesCompatibility = players => {
-    if (!players.filter(player => player.role.group === "werewolves").length) {
+    if (!players.filter(({ side }) => side.current === "werewolves").length) {
         throw generateError("NO_WEREWOLF_IN_GAME_COMPOSITION", "No player has the `werewolf` role in game composition.");
-    } else if (!players.filter(player => player.role.group === "villagers").length) {
+    } else if (!players.filter(({ side }) => side.current === "villagers").length) {
         throw generateError("NO_VILLAGER_IN_GAME_COMPOSITION", "No player has the `villager` role in game composition.");
+    } else if (getPlayerWithRole("two-sisters", { players }) &&
+        players.filter(({ role }) => role.current === "two-sisters").length !== 2) {
+        throw generateError("SISTERS_MUST_BE_TWO", `There must be exactly two sisters in game composition if at least one is chosen by a player.`);
+    } else if (getPlayerWithRole("three-brothers", { players }) &&
+        players.filter(({ role }) => role.current === "three-brothers").length !== 3) {
+        throw generateError("BROTHERS_MUST_BE_THREE", `There must be exactly three brothers in game composition if at least one is chosen by a player.`);
     }
 };
 
 exports.fillPlayersData = players => {
     for (const player of players) {
         player.name = filterOutHTMLTags(player.name);
-        const role = getPlayerRoles().find(playerRole => playerRole.name === player.role);
-        player.role = { original: role.name, current: role.name, group: role.group };
+        const role = getRoles().find(playerRole => playerRole.name === player.role);
+        player.role = { original: role.name, current: role.name };
+        player.side = { original: role.side, current: role.side };
     }
 };
 
@@ -79,7 +90,8 @@ exports.checkAndFillDataBeforeCreate = async data => {
     this.fillPlayersData(data.players);
     this.checkRolesCompatibility(data.players);
     await this.checkUserCurrentGames(data.gameMaster);
-    this.fillFirstWaiting(data);
+    this.fillTickData(data);
+    await this.fillWaitingQueueWithNightActions(data);
 };
 
 exports.create = async(data, options = {}) => {
@@ -157,17 +169,31 @@ exports.assignRoleToPlayers = (players, roles) => {
     return players;
 };
 
-exports.getVillagerRoles = async(players, werewolfRoles) => {
+exports.filterAvailablePowerfulVillagerRoles = (roles, players, leftToPick) => roles.filter(role => role.side === "villagers" &&
+    role.name !== "villager" && (!role.recommendedMinPlayers || role.recommendedMinPlayers <= players.length) &&
+    (!role.minInGame || role.minInGame <= leftToPick));
+
+exports.getVillagerRoles = (players, werewolfRoles) => {
     const villagerRoles = [];
     const villagerCount = players.length - werewolfRoles.length;
-    const availablePowerfulVillagerRoles = await getPlayerRoles().filter(role => role.group === "villagers" && role.name !== "villager");
-    const villagerRole = await getPlayerRoles().find(role => role.name === "villager");
+    const villagerRole = getRoles().find(role => role.name === "villager");
+    let availablePowerfulVillagerRoles = getRoles();
     for (let i = 0; i < villagerCount; i++) {
+        const leftToPick = villagerCount - i;
+        availablePowerfulVillagerRoles = this.filterAvailablePowerfulVillagerRoles(availablePowerfulVillagerRoles, players, leftToPick);
         if (!availablePowerfulVillagerRoles.length) {
             villagerRoles.push(JSON.parse(JSON.stringify(villagerRole)));
             villagerRole.maxInGame--;
         } else {
-            villagerRoles.push(this.pickRandomRole(availablePowerfulVillagerRoles));
+            const randomRole = this.pickRandomRole(availablePowerfulVillagerRoles);
+            if (randomRole.minInGame) {
+                for (let j = 0; j < randomRole.minInGame; j++) {
+                    villagerRoles.push(randomRole);
+                }
+                i += randomRole.minInGame - 1;
+            } else {
+                villagerRoles.push(randomRole);
+            }
         }
     }
     return villagerRoles;
@@ -177,7 +203,7 @@ exports.pickRandomRole = roles => {
     const idx = Math.floor(Math.random() * roles.length);
     const role = JSON.parse(JSON.stringify(roles[idx]));
     roles[idx].maxInGame--;
-    if (!roles[idx].maxInGame) {
+    if (!roles[idx].maxInGame || roles[idx].minInGame) {
         roles.splice(idx, 1);
     }
     return role;
@@ -200,19 +226,19 @@ exports.getWerewolfCount = players => {
 exports.getWerewolfRoles = players => {
     const werewolfRoles = [];
     const werewolfCount = this.getWerewolfCount(players);
-    const availableWerewolfRoles = getPlayerRoles().filter(role => role.group === "werewolves");
+    const availableWerewolfRoles = getRoles().filter(role => role.side === "werewolves");
     for (let i = 0; i < werewolfCount; i++) {
         werewolfRoles.push(this.pickRandomRole(availableWerewolfRoles));
     }
     return werewolfRoles;
 };
 
-exports.getGameRepartition = async(req, res) => {
+exports.getGameRepartition = (req, res) => {
     try {
         const { query } = checkRequestData(req);
         this.checkUniqueNameInPlayers(query.players);
-        const werewolfRoles = await this.getWerewolfRoles(query.players);
-        const villagerRoles = await this.getVillagerRoles(query.players, werewolfRoles);
+        const werewolfRoles = this.getWerewolfRoles(query.players);
+        const villagerRoles = this.getVillagerRoles(query.players, werewolfRoles);
         this.assignRoleToPlayers(query.players, [...villagerRoles, ...werewolfRoles]);
         res.status(200).json({ players: query.players });
     } catch (e) {
@@ -241,7 +267,7 @@ exports.postGame = async(req, res) => {
         const { body } = checkRequestData(req);
         const game = await this.create({
             gameMaster: new mongoose.Types.ObjectId(req.user._id),
-            players: body.players,
+            ...body,
         });
         res.status(200).json(game);
     } catch (e) {
@@ -272,10 +298,12 @@ exports.checkGameWinners = game => {
     if (isGameDone(game)) {
         if (areAllPlayersDead(game)) {
             game.won = { by: null };
+        } else if (areLoversTheOnlyAlive(game)) {
+            game.won = { by: "lovers", players: getPlayersWithAttribute("in-love", game) };
         } else if (!isVillagerSideAlive(game)) {
-            game.won = { by: "werewolves", players: game.players.filter(player => player.role.group === "werewolves") };
+            game.won = { by: "werewolves", players: getPlayersWithSide("werewolves", game) };
         } else if (!isWerewolfSideAlive(game)) {
-            game.won = { by: "villagers", players: game.players.filter(player => player.role.group === "villagers") };
+            game.won = { by: "villagers", players: getPlayersWithSide("villagers", game) };
         }
         game.status = "done";
     }
@@ -304,9 +332,13 @@ exports.fillWaitingQueueWithDayActions = async game => {
         { attribute: "eaten", trigger: Player.eaten },
         { attribute: "drank-death-potion", trigger: Player.drankDeathPotion },
     ];
-    for (const { attribute, trigger } of playerAttributeMethods) {
-        if (getPlayerWithAttribute(attribute, game)) {
-            trigger(game);
+    const alivePlayers = getAlivePlayers(game);
+    for (const player of alivePlayers) {
+        for (const { attribute, trigger } of playerAttributeMethods) {
+            const playerAttribute = getPlayerAttribute(player, attribute);
+            if (playerAttribute) {
+                trigger(game, playerAttribute);
+            }
         }
     }
     this.purgeAttributesAfterSunRising(game);
@@ -315,69 +347,93 @@ exports.fillWaitingQueueWithDayActions = async game => {
     }
 };
 
-exports.isSourceAvailableInPlayers = (players, source) => {
+exports.isGroupCallableDuringTheNight = (game, group) => {
+    if (group === "lovers" && getPlayerWithRole("cupid", game)) {
+        return true;
+    }
+    const players = getPlayersWithSide(group, game);
+    return game.tick === 1 ? !!players.length : !!players.length && players.some(({ isAlive }) => isAlive);
+};
+exports.areThreeBrothersCallableDuringTheNight = async game => {
+    const lastBrothersPlay = await GameHistory.getLastBrothersPlay(game._id);
+    const brotherPlayers = getPlayersWithRole("three-brothers", game);
+    const turnsSinceLastBrothersPlay = game.turn - lastBrothersPlay.turn;
+    return brotherPlayers.filter(brother => brother.isAlive).length >= 2 &&
+        (!lastBrothersPlay || turnsSinceLastBrothersPlay >= game.options.brothersWakingUpInterval && game.options.brothersWakingUpInterval);
+};
+exports.areTwoSistersCallableDuringTheNight = async game => {
+    const lastSistersPlay = await GameHistory.getLastSistersPlay(game._id);
+    const sisterPlayers = getPlayersWithRole("two-sisters", game);
+    const turnsSinceLastSistersPlay = game.turn - lastSistersPlay.turn;
+    return sisterPlayers.every(({ isAlive }) => isAlive) &&
+        (!lastSistersPlay || turnsSinceLastSistersPlay >= game.options.sistersWakingUpInterval && game.options.sistersWakingUpInterval);
+};
+
+exports.isRoleCallableDuringTheNight = (game, role) => {
+    const player = getPlayerWithRole(role, game);
+    if (!player || game.tick === 1) {
+        return !!player;
+    }
+    if (role === "two-sisters") {
+        return this.areTwoSistersCallableDuringTheNight(game);
+    } else if (role === "three-brothers") {
+        return this.areThreeBrothersCallableDuringTheNight(game);
+    } else if (role === "big-bad-wolf") {
+        return player.isAlive && areAllWerewolvesAlive(game);
+    }
+    return game.tick === 1 ? !!player : !!player && player.isAlive;
+};
+
+exports.isSourceCallableDuringTheNight = (game, source) => {
     if (source === "all" || source === "sheriff") {
         return true;
     }
-    const sourceType = groupNames.includes(source) ? "group" : "role";
-    for (const player of players) {
-        if (sourceType === "group" && player.role.group === source && player.isAlive ||
-            sourceType === "role" && player.role.current === source && player.isAlive) {
-            return true;
-        }
-    }
-    return false;
+    const sourceType = getGroupNames().includes(source) ? "group" : "role";
+    return sourceType === "group" ? this.isGroupCallableDuringTheNight(game, source) : this.isRoleCallableDuringTheNight(game, source);
 };
 
-exports.fillWaitingQueueWithNightActions = async(game, endOfDay = false) => {
-    const actionsOrder = game.turn === 1 ? [...turnPreNightActionsOrder, ...turnNightActionsOrder] : [...turnNightActionsOrder];
-    if (endOfDay) {
-        for (let i = 0; i < actionsOrder.length; i++) {
-            if (this.isSourceAvailableInPlayers(game.players, actionsOrder[i].source)) {
-                const nextGameNightAction = actionsOrder[i];
-                return game.waiting.push({ for: nextGameNightAction.source, to: nextGameNightAction.action });
-            }
-        }
-    } else {
-        let actionFound = false;
-        const lastNightPlay = await GameHistory.getLastNightPlay(game._id);
-        for (let i = 0; i < actionsOrder.length; i++) {
-            if (lastNightPlay && lastNightPlay.play.source === actionsOrder[i].source && lastNightPlay.play.action === actionsOrder[i].action) {
-                actionFound = true;
-            } else if (actionFound && this.isSourceAvailableInPlayers(game.players, actionsOrder[i].source)) {
-                const nextGameNightAction = actionsOrder[i];
-                return game.waiting.push({ for: nextGameNightAction.source, to: nextGameNightAction.action });
-            }
+exports.fillWaitingQueueWithNightActions = async game => {
+    const actionsOrder = game.tick === 1 ? getGameTurNightActionsOrder() : getGameTurNightActionsOrder().filter(action => !action.isFirstNightOnly);
+    if (!game.waiting) {
+        game.waiting = [];
+    }
+    for (const { source, action } of actionsOrder) {
+        if (await this.isSourceCallableDuringTheNight(game, source)) {
+            game.waiting.push({ for: source, to: action });
         }
     }
 };
 
 exports.fillWaitingQueue = async game => {
     if (game.phase === "night") {
-        await this.fillWaitingQueueWithNightActions(game);
-        if (!game.waiting || !game.waiting.length) {
-            game.phase = "day";
-            return this.fillWaitingQueueWithDayActions(game);
-        }
+        game.phase = "day";
+        return this.fillWaitingQueueWithDayActions(game);
     } else if (game.phase === "day") {
         await this.fillWaitingQueueWithDayActions(game);
         if (!game.waiting || !game.waiting.length) {
             game.phase = "night";
             game.turn++;
-            return this.fillWaitingQueueWithNightActions(game, true);
+            return this.fillWaitingQueueWithNightActions(game);
         }
     }
 };
 
 exports.generatePlayMethods = () => ({
-    all: Player.allPlay,
-    seer: Player.seerPlays,
-    witch: Player.witchPlays,
-    guard: Player.guardPlays,
-    raven: Player.ravenPlays,
-    hunter: Player.hunterPlays,
-    werewolves: Player.werewolvesPlay,
-    sheriff: Player.sheriffPlays,
+    "all": Player.allPlay,
+    "seer": Player.seerPlays,
+    "witch": Player.witchPlays,
+    "guard": Player.guardPlays,
+    "raven": Player.ravenPlays,
+    "hunter": Player.hunterPlays,
+    "werewolves": Player.werewolvesPlay,
+    "sheriff": Player.sheriffPlays,
+    "cupid": Player.cupidPlays,
+    "lovers": () => undefined,
+    "two-sisters": () => undefined,
+    "three-brothers": () => undefined,
+    "wild-child": Player.wildChildPlays,
+    "dog-wolf": Player.dogWolfPlays,
+    "big-bad-wolf": Player.bigBadWolfPlays,
 });
 
 exports.generateGameHistoryEntry = (game, play) => ({
@@ -444,15 +500,11 @@ exports.resetGame = async(req, res) => {
         await this.checkGameBeforeReset(params.id);
         await GameHistory.deleteMany({ gameId: params.id });
         let game = await this.findOne({ _id: params.id });
-        const players = game.players.map(player => ({ name: player.name, role: player.role.original }));
-        await this.fillPlayersData(players);
-        game = await this.findOneAndUpdate({ _id: params.id }, {
-            players,
-            turn: 1,
-            phase: "night",
-            tick: 1,
-            waiting: [{ for: "all", to: "elect-sheriff" }],
-        });
+        const resetData = { players: game.players.map(player => ({ name: player.name, role: player.role.original })) };
+        await this.fillPlayersData(resetData.players);
+        this.fillTickData(resetData);
+        await this.fillWaitingQueueWithNightActions(resetData);
+        game = await this.findOneAndUpdate({ _id: params.id }, resetData);
         res.status(200).json(game);
     } catch (e) {
         sendError(res, e);