+++ src/helpers/functions/Player.js
@@ -1,14 +1,17 @@
 const { playerAttributes, murderedPossibilities, playerActions } = require("../constants/Player");
 
-exports.canBeEaten = player => !this.hasAttribute(player, "drank-life-potion") && (!this.hasAttribute(player, "protected") ||
+exports.canBeEaten = player => !this.doesPlayerHaveAttribute(player, "drank-life-potion") && (!this.doesPlayerHaveAttribute(player, "protected") ||
                                 player.role.current === "little-girl");
 
-exports.hasAttribute = ({ attributes }, attributeName) => attributes && attributes.findIndex(({ attribute }) => attribute === attributeName) !== -1;
+exports.doesPlayerHaveAttribute = ({ attributes }, attributeName) => attributes &&
+    attributes.findIndex(({ attribute }) => attribute === attributeName) !== -1;
 
-exports.getPlayerAttributes = () => JSON.parse(JSON.stringify(playerAttributes));
+exports.getAttributes = () => JSON.parse(JSON.stringify(playerAttributes));
 
-exports.getPlayerAttribute = attribute => this.getPlayerAttributes().find(playerAttribute => playerAttribute.attribute === attribute);
+exports.getAttribute = attribute => this.getAttributes().find(playerAttribute => playerAttribute.attribute === attribute);
 
 exports.getPlayerActions = () => JSON.parse(JSON.stringify(playerActions));
 
 exports.getPlayerMurderedPossibilities = () => JSON.parse(JSON.stringify(murderedPossibilities));
+
+exports.getPlayerAttribute = ({ attributes }, attributeName) => attributes && attributes.find(({ attribute }) => attribute === attributeName);